Refactor piece receiving method.
- move piece receiving funtionality from farmer to networking crate to enable its reuse
