+++ crates/subspace-farmer-components/Cargo.toml
@@ -27,6 +27,7 @@ serde = { version = "1.0.147", features = ["derive"] }
 static_assertions = "1.1.0"
 subspace-solving = { version = "0.1.0", path = "../subspace-solving" }
 subspace-core-primitives = { version = "0.1.0", path = "../subspace-core-primitives" }
+subspace-networking = { version = "0.1.0", path = "../subspace-networking" }
 subspace-verification = { version = "0.1.0", path = "../subspace-verification" }
 thiserror = "1.0.32"
 tracing = "0.1.37"
