+++ crates/subspace-networking/src/lib.rs
@@ -60,5 +60,6 @@ pub use request_handlers::pieces_by_range::{
 };
 pub use utils::deconstruct_record_key;
 pub use utils::multihash::ToMultihash;
+pub use utils::piece_receiver::{PieceProvider, PieceReceiver, PieceValidator};
 pub use utils::prometheus::start_prometheus_metrics_server;
 pub use utils::record_binary_heap::RecordBinaryHeap;
