+++ test/subspace-test-client/Cargo.toml
@@ -31,6 +31,7 @@ subspace-archiving = { path = "../../crates/subspace-archiving" }
 subspace-core-primitives = { path = "../../crates/subspace-core-primitives" }
 subspace-runtime-primitives = { path = "../../crates/subspace-runtime-primitives" }
 subspace-farmer-components = { path = "../../crates/subspace-farmer-components" }
+subspace-networking = { version = "0.1.0", path = "../../crates/subspace-networking" }
 subspace-service = { path = "../../crates/subspace-service" }
 subspace-solving = { path = "../../crates/subspace-solving" }
 subspace-test-runtime = { version = "0.1.0", features = ["do-not-enforce-cost-of-storage"], path = "../subspace-test-runtime" }
